pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "presentations_module"
version = "0.2.16"
dependencies = [
  "playwright==1.57.0",
  "pymongo==4.16.0",
  "python-dotenv==1.2.1",
  "aioboto3>=13.0",
  "paramiko>=3.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.24",
  "pylint>=3.0",
]
